\name{add.posterior.density}
\alias{add.posterior.density}
\title{Add posterior density.}
\usage{
  add.posterior.density(posterior.density, base.plot =
  NULL)
}
\arguments{
  \item{posterior.density}{}

  \item{base.plot}{a ggplot object (grob), created by
  \code{\link{create.base.plot}} if none provided}
}
\value{
  plot with posterior density line added
}
\description{
  Add posterior density to a plot
}
\author{
  David LeBauer
}

