% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alberta_flag.R, R/utils.R
\name{add_alberta_flag}
\alias{add_alberta_flag}
\title{Add Alberta flag to a site data frame using precise boundaries}
\usage{
add_alberta_flag(dat, lat_col = "latitude", lon_col = "longitude")

add_alberta_flag(dat, lat_col = "latitude", lon_col = "longitude")
}
\arguments{
\item{dat}{Data frame with latitude and longitude columns}

\item{lat_col}{Name of latitude column (default: "latitude")}

\item{lon_col}{Name of longitude column (default: "longitude")}
}
\value{
Data frame with in_alberta column ("in" or "out")

Data frame with in_alberta column ("in" or "out")
}
\description{
Uses rnaturalearth and sf to determine if each site is in Alberta.

Uses rnaturalearth and sf to determine if each site is in Alberta.
}
