% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arrhenius.scaling}
\alias{arrhenius.scaling}
\title{Arrhenius scaling}
\usage{
arrhenius.scaling(observed.value, old.temp, new.temp = 25)
}
\arguments{
\item{observed.value}{observed value of temperature dependent trait, e.g. Vcmax, root respiration rate}

\item{old.temp}{temperature at which measurement was taken or previously scaled to}

\item{new.temp}{temperature to be scaled to, default = 25 C}
}
\value{
numeric value at reference temperature
}
\description{
Scale temperature dependent trait from measurement temperature to reference temperature
}
\author{
unknown
}

