% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bibtexify}
\alias{bibtexify}
\title{bibtexify}
\usage{
bibtexify(author, year, title)
}
\arguments{
\item{author}{name of first author}

\item{year}{year of publication}

\item{title}{manuscript title}
}
\value{
bibtex citation
}
\description{
Converts author year title to bibtex \code{author1999abc} format
}
\author{
unknown
}
