% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{string}
}
\value{
x, capitalized
}
\description{
Capitalize a string
}
\author{
David LeBauer
}

