% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf2date.R
\name{cf2datetime}
\alias{cf2datetime}
\title{Convert CF-style date-time to POSIXct date-time}
\usage{
cf2datetime(value, unit, tz = "UTC")
}
\arguments{
\item{value}{Numeric value of CF date-time}

\item{unit}{CF style unit (e.g. "days since 2010-01-01")}

\item{tz}{Time zone of result (default = "UTC")}
}
\value{
POSIXct datetime
}
\description{
Convert CF-style date-time to POSIXct date-time
}
\examples{
cf2datetime(5, "days since 1981-01-01")
cf2datetime(27, "minutes since 1963-01-03 12:00:00 -05:00")
# no leap year
cf2datetime(365, "days since 1999-01-01")
# leap year
cf2datetime(365, "days since 2000-01-01 12:00:00 -05:00")
}
\author{
Alexey Shiklomanov
}
