\name{check.settings}
\alias{check.settings}
\title{Sanity checks. Checks the settings file to make sure expected fields exist.}
\usage{
  check.settings(settings)
}
\description{
  Expected fields are: - pfts with at least one pft defined
  - database - model - run with the following fields --
  site with id -- host with name
}

