\name{clear.scratch}
\alias{clear.scratch}
\title{Clear EBI-CLUSTER worker node local scratch directories of old PEcAn output}
\value{
  nothing
}
\description{
  Removes previous model run output from worker node local
  scratch directories on EBI-CLUSTER
}
\examples{
\dontrun{
clear.scratch(settings)
}
}
\author{
  Shawn Serbin
}

