% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{conservation_level}
\alias{conservation_level}
\title{conservation_level}
\usage{
conservation_level(consensus)
}
\arguments{
\item{consensus}{String of the consensus multiple local alignment.}
}
\value{
The conservation level of the consensus string.
}
\description{
computes de the conservation level from consensus string.
}
