% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert.expr}
\alias{convert.expr}
\title{convert.expr}
\usage{
convert.expr(expr)
}
\arguments{
\item{expr}{expression string}
}
\value{
variables variable names
}
\description{
Convert expression to variable names
}
\author{
Istem Fer
}

