% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.input.R
\name{convert.input}
\alias{convert.input}
\title{convert.input}
\usage{
convert.input(input.id, outfolder, formatname, mimetype, site.id, start_date,
  end_date, pkg, fcn, con = con, host, browndog, write = TRUE, format.vars,
  overwrite = FALSE, exact.dates = FALSE, allow.conflicting.dates = TRUE,
  insert.new.file = FALSE, pattern = NULL, ...)
}
\description{
Convert input by applying fcn and insert new record into database
}
\author{
Betsy Cowdery, Michael Dietze, Ankur Desai, Tony Gardella
}
