% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.utils.R
\name{counter}
\alias{counter}
\title{counter}
\usage{
counter(cnt)
}
\arguments{
\item{cnt}{}
}
\value{
updated value of cnt to global environment
}
\description{
Counter function for writing configs
}

