% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create.base.plot}
\alias{create.base.plot}
\title{Create Base Plot}
\usage{
create.base.plot()
}
\value{
empty ggplot object
}
\description{
Creates empty ggplot object
}
\details{
An empty base plot to which layers created by other functions
(\code{\link{plot.data}}, \code{\link{plot.prior.density}},
\code{\link{plot.posterior.density}}) can be added.
}
\author{
David LeBauer
}

