\name{create.density.df}
\alias{create.density.df}
\title{Create Density Data Frame from Sample}
\usage{
  create.density.df(samps = NULL, zero.bounded = FALSE,
    distribution = NULL)
}
\arguments{
  \item{sample}{}

  \item{...}{additional arguments to density}
}
\value{
  data frame with x and y
}
\description{
  Returns a data frame from \link{stats::density} function
}
\author{
  David LeBauer
}

