% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf2date.R
\name{datetime2cf}
\alias{datetime2cf}
\title{Convert POSIXct date-time to CF-style date-time}
\usage{
datetime2cf(datetime, unit, ...)
}
\arguments{
\item{datetime}{POSIXct datetime, or object that can be to POSIXct
via \code{as.POSIXct}}

\item{unit}{Target CF-style unit (e.g. "days since 2010-01-01")}

\item{...}{Additional arguments to \code{as.POSIXct}. A common one is
\code{tz} for time-zone (e.g. \code{tz = "UTC"}).}
}
\value{
Numeric value of date-time in target CF unit
}
\description{
Convert POSIXct date-time to CF-style date-time
}
\examples{
datetime2cf("1990-10-05", "days since 1990-01-01", tz = "UTC")
}
