% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_year.R
\name{days_in_year}
\alias{days_in_year}
\title{Number if days in given year}
\usage{
days_in_year(year)
}
\arguments{
\item{year}{Calendar year}
}
\description{
Accounts for leap years.
}
\author{
Alexey Shiklomanov
}
