% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distn.stats.R
\name{distn.stats}
\alias{distn.stats}
\title{Distribution Stats}
\usage{
distn.stats(distn, a, b)
}
\arguments{
\item{distn}{named distribution, one of 'beta', 'exp', 'f', 'gamma', 'lnorm', 'norm', 't',}

\item{a}{numeric; first parameter of \code{distn}}

\item{b}{numeric; second parameter of \code{distn}}
}
\value{
vector with mean and standard deviation
}
\description{
Implementation of standard equations used to calculate mean and sd for a variety of
named distributions different
}
\examples{
distn.stats("norm", 0, 1)
}
\author{
David LeBauer
}
