% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distn.stats.R
\name{distn.table.stats}
\alias{distn.table.stats}
\title{Helper function for computing summary statistics of a parametric distribution}
\usage{
distn.table.stats(distns)
}
\arguments{
\item{distns}{table of distributions; see examples}
}
\value{
named vector of mean and SD
}
\description{
return mean and standard deviation of a distribution for each distribution in a table with \code{colnames = c('distn', 'a', 'b')},
e.g. in a table of priors
}
\author{
David LeBauer
}
