% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.conversions.R
\name{do.conversions}
\alias{do.conversions}
\title{do.conversions}
\usage{
do.conversions(settings, overwrite.met = FALSE, overwrite.fia = FALSE,
  overwrite.ic = FALSE)
}
\description{
Input conversion workflow
}
\author{
Ryan Kelly, Rob Kooper, Betsy Cowdery, Istem Fer
}
