% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download.file}
\alias{download.file}
\title{download.file}
\usage{
download.file(url, destfile, method = "ncftpget")
}
\arguments{
\item{url}{complete URL for file download}

\item{destfile}{destination file name}

\item{method}{Method of file retrieval.  Currently the only option is ncftpget}
}
\description{
Simple function to use ncftpget for FTP downloads behind a firewall.
Requires ncftpget and a properly formatted config file in the users 
home directory
}
\author{
Shawn Serbin
}

