% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.url.R
\name{download.url}
\alias{download.url}
\title{Try and download a file.}
\usage{
download.url(url, file, timeout = 600, .opts = list(), retry = TRUE)
}
\arguments{
\item{url}{the url of the file to download}

\item{file}{the filename}

\item{timeout}{number of seconds to wait for file (default 600)}

\item{.opts}{list of options for curl, for example to download from a
protected site use list(userpwd=userpass, httpauth = 1L)}

\item{retry}{retry if url not found yet, this is used by Brown Dog}
}
\value{
returns name of file if successful or NA if not.
}
\description{
This will download a file, if retry is set and 404 is returned it will
wait until the file is available. If the file is still not available
after timeout tries, it will return NA. If the file is downloaded
it will return the name of the file
}
\examples{
\dontrun{
download.url("http://localhost/", index.html)
}
}
