% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Simple function to use ncftpget for FTP downloads behind a firewall.}
\usage{
download_file(url, filename, method)
}
\arguments{
\item{url}{complete URL for file download}

\item{filename}{destination file name}

\item{method}{Method of file retrieval. Can set this using the \verb{options(download.ftp.method=[method])} in your Rprofile.
example options(download.ftp.method="ncftpget")}
}
\description{
Requires ncftpget and a properly formatted config file in the users
home directory
}
\examples{
\dontrun{
download_file("http://lib.stat.cmu.edu/datasets/csb/ch11b.txt", "~/test.download.txt")

download_file(
  "
  ftp://ftp.cdc.noaa.gov/Datasets/NARR/monolevel/pres.sfc.2000.nc",
  "~/pres.sfc.2000.nc"
)
}

}
\author{
Shawn Serbin, Rob Kooper
}
