% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{extract_jaspar_metadata}
\alias{extract_jaspar_metadata}
\title{extract_jaspar_metadata}
\usage{
extract_jaspar_metadata(version)
}
\arguments{
\item{version}{This is 2020 or 2022.}
}
\value{
The metadata of all motifs in the JASPAR repository.
}
\description{
extract JASPAR 2020 or 2022 metadata of all motifs.
}
