% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_model_netcdf.R
\name{extract_nc_sda}
\alias{extract_nc_sda}
\title{Extract netCDF file by site.id, time window, and variable name.}
\usage{
extract_nc_sda(site.id, start.date, end.date, var.name, nc.path)
}
\arguments{
\item{site.id}{numeric or character: identification of the site.}

\item{start.date}{date in YYYY-MM-DD format: start date of the requested time window.}

\item{end.date}{date in YYYY-MM-DD format: end date of the requested time window.}

\item{var.name}{character: variable name.}

\item{nc.path}{character: physical path to the target netCDF file.}
}
\value{
list: a list contains requested array, time steps, site id, variable name, and ensemble size.
}
\description{
Extract netCDF file by site.id, time window, and variable name.
}
\details{
The function is only tested for netCDF files generated by the \code{nc_merge_all_sites_by_year} function.
}
\author{
Dongchen Zhang
}
