% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{generate_ANR2}
\alias{generate_ANR2}
\title{generate_ANR2}
\usage{
generate_ANR2(clevel, k, n = 20000, t = 105, maxperseq = 5)
}
\arguments{
\item{clevel}{Conservation level.}

\item{k}{Size of motif to be inserted.}

\item{n}{Number of sequences.}

\item{t}{Size of each sequence.}

\item{maxperseq}{Max number of motifs to generate per sequence. Using only ANR distribution.}
}
\value{
Synthetic ANR dataset.
}
\description{
Generate synthetic ANR dataset.
}
