% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{generate_OOPS}
\alias{generate_OOPS}
\title{generate_OOPS}
\usage{
generate_OOPS(clevel, k, n = 20000, t = 105, pos = 50)
}
\arguments{
\item{clevel}{Conservation level.}

\item{k}{Size of motif to be inserted.}

\item{n}{Number of sequences.}

\item{t}{Size of each sequence.}

\item{pos}{Position of the motif in the sequences.}
}
\value{
Synthetic OOPS dataset.
}
\description{
Generate synthetic OOPS dataset.
}
