% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{generate_ZOOPS}
\alias{generate_ZOOPS}
\title{generate_ZOOPS}
\usage{
generate_ZOOPS(clevel, k, n = 20000, t = 105, pos = 50, noise = 0.05)
}
\arguments{
\item{clevel}{Conservation level.}

\item{k}{Size of motif to be inserted.}

\item{n}{Number of sequences.}

\item{t}{Size of each sequence.}

\item{pos}{Position of the motif in the sequences.}

\item{noise}{Level of noise.}
}
\value{
Synthetic ZOOPS dataset.
}
\description{
Generate synthetic ZOOPS dataset.
}
