% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ensemble.inputs.R
\name{get.ensemble.inputs}
\alias{get.ensemble.inputs}
\title{get.ensemble.inputs}
\usage{
get.ensemble.inputs(settings, ens = 1)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{ens}{ensemble number. default = 1}
}
\value{
find correct ensemble inputs
}
\description{
Splits climate met for SIPNET
}
\author{
Mike Dietze and Ann Raiho
}
