% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.model.output.R
\name{get.model.output}
\alias{get.model.output}
\title{Retrieve model output}
\usage{
get.model.output(model, settings)
}
\arguments{
\item{model}{the ecosystem model run}

\item{settings}{list of PEcAn settings.}
}
\description{
This function retrieves model output for further analyses
}
\examples{
\dontrun{
get.model.output(model)
get.model.output('ED2')
}

}
\author{
Michael Dietze, Shawn Serbin, David LeBauer
}
