% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.parameter.stat}
\alias{get.parameter.stat}
\title{Get Parameter Statistics}
\usage{
get.parameter.stat(mcmc.summary, parameter)
}
\arguments{
\item{mcmc.summary}{probably produced by \code{\link[coda]{summary.mcmc}}}

\item{parameter}{name of parameter to extract, as character}
}
\value{
table with parameter statistics
}
\description{
Gets statistics for LaTeX - formatted table
}
\examples{
\dontrun{
get.parameter.stat(mcmc.summaries[[1]], "beta.o")
}
}
\author{
David LeBauer
}
