% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.config.utils.R
\name{get.quantiles}
\alias{get.quantiles}
\title{Get Quantiles}
\usage{
get.quantiles(quantiles.tag)
}
\arguments{
\item{quantiles.tag}{specifies tag used to specify quantiles}
}
\value{
vector of quantiles
}
\description{
Returns a vector of quantiles specified by a given <quantiles> xml tag
}
\author{
David LeBauer
}

