\name{get.results}
\alias{get.results}
\title{Generate model output for PEcAn analyses}
<<<<<<< HEAD
\usage{
  get.results(pecandir, model, settings = settings)
}
=======
>>>>>>> ef9572c0bfd74f64f90e160d3d29f5e3f933b0b6
\arguments{
  \item{pecandir}{output directory}

  \item{model}{name of model being used}

  \item{settings}{list, read from settings file (xml) using
  \code{\link{read.settings}}}
}
\description{
  Reads model output and runs sensitivity and ensemble
  analyses
}
\details{
  Output is placed in model output directory
  (settings$run$host$outdir).
}
\author{
  Shawn Serbin, David LeBauer, Mike Dietze

  David LeBauer, Shawn Serbin, Mike Dietze
}

