% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.new.site.R
\name{get.site.info}
\alias{get.site.info}
\title{Get New Site Info}
\usage{
get.site.info(site, con)
}
\arguments{
\item{site}{a dataframe with site information on lat, lon and time_zone}

\item{con}{Database connection object}
}
\value{
a dataframe with new site information on lat, lon and time_zone
}
\description{
Get new site info using provided site information
}
\examples{
get.site.info(site = data.frame(lat = 40.1, lon = -88.2, time_zone = "America/Chicago"), con = con)
}
\author{
Abhinav Pandey
}
