% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.stats.mcmc}
\alias{get.stats.mcmc}
\title{Further summarizes output from summary.mcmc}
\usage{
get.stats.mcmc(mcmc.summary, sample.size)
}
\arguments{
\item{mcmc.summary}{probably produced by \code{\link[coda]{summary.mcmc}}}

\item{sample.size}{passed as 'n' in returned list}
}
\value{
list with summary statistics for parameters in an MCMC chain
}
\description{
Further summarizes output from summary.mcmc
}
\author{
David LeBauer
}
