% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grid2netcdf}
\alias{grid2netcdf}
\title{grid2netcdf}
\usage{
grid2netcdf(gdata, date = "9999-09-09", outfile = "out.nc")
}
\arguments{
\item{grid.data}{}
}
\value{
writes netCDF file
}
\description{
Write gridded data to netcdf file
}
\author{
David LeBauer
}

