% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_assertions.r
\name{has_unique_names}
\alias{has_unique_names}
\title{Check that R object has full set of names}
\usage{
has_unique_names(x)
}
\arguments{
\item{x}{An R object}
}
\value{
logical, \code{TRUE} if object has full set of unique names.
}
\description{
Check if an R object has a unique set of names. Every element of the object
must have a non-NA, non-empty string name.
}
\author{
Andrew Roberts
}
