% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{input.ens.gen}
\alias{input.ens.gen}
\title{Function for reurning the met ensuble based on sampling method, parent or etc}
\usage{
input.ens.gen(settings, input, method = "sampling", parenids = NULL, ...)
}
\arguments{
\item{settings}{list of PEcAn settings}

\item{method}{Method for sampling - For now looping or sampling with replacement is implemented}

\item{parenids}{If met's ids are read from a parent what are the ids}

\item{...}{}
}
\description{
Function for reurning the met ensuble based on sampling method, parent or etc
}
