% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{iqr}
\alias{iqr}
\title{Interquartile range}
\usage{
iqr(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric vector of length 2, with the 25th and 75th quantiles of input vector x.
}
\description{
Calculate interquartile range
}
\details{
Calculates the 25th and 75th quantiles given a vector x; used in function \link{dhist}.
}

