% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remote.R
\name{is.localhost}
\alias{is.localhost}
\title{Check if local host}
\usage{
is.localhost(host)
}
\arguments{
\item{host}{the hostname to be checked}
}
\value{
true if the host is the local host name
}
\description{
Check if host is local
}
\details{
Given the hostname is this the localhost. This returns true if either
the value is localhost, or the value is the same as the fqdn.
}
\examples{
is.localhost(fqdn())
}
\author{
Rob Kooper
}

