% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_assertions.r
\name{is_lower_tri}
\alias{is_lower_tri}
\title{Check if a matrix is lower triangular}
\usage{
is_lower_tri(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
\code{TRUE} if \code{m} is lower triangular, else \code{FALSE}.
}
\description{
A lower triangular matrix is one in which all entries strictly above the
main diagonal are zero. Note that this function will run for non-square
matrices, but it is primarily designed for square matrices.
}
\seealso{
\code{\link{is_upper_tri}}
}
\author{
Andrew Roberts
}
