% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_assertions.r
\name{is_nonneg_integer_vector}
\alias{is_nonneg_integer_vector}
\title{Check for integerish vector with non-negative entries.}
\usage{
is_nonneg_integer_vector(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Check for integerish vector with non-negative entries.
}
\author{
Andrew Roberts
}
