% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_assertions.r
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Check if a matrix is positive definite, and return Cholesky factor}
\usage{
is_positive_definite(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
Logical, \code{TRUE} if \code{chol(m)} does not throw an error.
If \code{TRUE}, the \code{chol_upper} attribute is defined and contains
the upper Cholesky factor.
}
\description{
Defines "positive definite" in the numerical sense of being able to run
\code{chol(m)} without an error. To avoid wasting this computation, returns
the result of \code{chol(m)} (the upper Cholesky factor) as an attribute
\code{chol_upper} of the returned logical value. This attribute is only set
when Cholesky decomposition does not throw an error.
}
\author{
Andrew Roberts
}
