% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_assertions.r
\name{is_raw_scalar}
\alias{is_raw_scalar}
\title{Check if an R object is an atomic scalar}
\usage{
is_raw_scalar(x)
}
\arguments{
\item{x}{An R object}
}
\value{
logical, \code{TRUE} if a raw scalar.
}
\description{
Must be atomic and length one. In contrast to \code{assertthat::is.scalar()},
singleton arrays, scalars, and one-element lists are not considered
raw scalars.
}
\author{
Andrew Roberts
}
