\name{left.pad.zeros}
\alias{left.pad.zeros}
\title{Left Pad Zeros}
\usage{
  left.pad.zeros(num, digits = 5)
}
\arguments{
  \item{num}{number to be padded (integer)}

  \item{digits}{number of digits to add}
}
\value{
  num with zeros to the left
}
\description{
  left padded by zeros up to a given number of digits.
}
\details{
  returns a string representing a given number
}
\author{
  Carl Davidson
}

