% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToXml.default}
\alias{listToXml.default}
\title{List to XML}
\usage{
\method{listToXml}{default}(item, tag)
}
\arguments{
\item{item}{}

\item{tag}{xml tag}
}
\value{
xmlNode
}
\description{
Convert List to XML
}
\details{
Can convert list or other object to an xml object using xmlNode
}
\author{
David LeBauer, Carl Davidson, Rob Kooper
}

