% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_local.R
\name{load_local}
\alias{load_local}
\title{Load an \code{RData} file into a list}
\usage{
load_local(file)
}
\arguments{
\item{file}{a (readable binary-mode) \link[base]{connection} or a character string
    giving the name of the file to load (when \link[base]{tilde expansion}
    is done).}
}
\value{
List, with names corresponding to object names in \code{file}
}
\description{
Instead of polluting the current environment, this allows you to
read an \code{RData} file into a list object of whatever name you choose.
}
\examples{
x <- 1:10
y <- 11:15
tmp <- tempfile()
save(x, y, file = tmp)
my_list <- load_local(tmp)
rm(tmp)
}
\author{
Alexey Shiklomanov
}
