% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_most_recent_data.R, R/utils.R
\name{load_most_recent_data}
\alias{load_most_recent_data}
\title{Load the Most Recent Data File Dynamically}
\usage{
load_most_recent_data(directory, file_prefix)

load_most_recent_data(directory, file_prefix)
}
\arguments{
\item{directory}{character. Path to the directory containing
the \code{.rData} files.}

\item{file_prefix}{character. Prefix of the file to search for
(e.g., "data_for_models_").}
}
\value{
Loads the data from the most recent \code{.rData} file into
the global environment and prints the name of the loaded file.

Loads the data from the most recent \code{.rData} file into
the global environment and prints the name of the loaded file.
}
\description{
This function dynamically loads the most recent \code{.rData} file
from a specified directory that matches a given file prefix and
follows the naming convention: \link{prefix}YYYY-MM-DD.rData.

This function dynamically loads the most recent \code{.rData} file
from a specified directory that matches a given file prefix and
follows the naming convention: \link{prefix}YYYY-MM-DD.rData.
}
