\name{log.info}
\alias{log.info}
\title{Prints an informational message.}
\usage{
  log.info(msg, ...)
}
\arguments{
  \item{msg}{the message that should be printed.}

  \item{...}{any additional text that should be printed.}
}
\description{
  This function will print an informational message.
}
\examples{
\dontrun{
log.info("PEcAn version 1.2")
}
}
\author{
  Rob Kooper
}

