\name{log.warn}
\alias{log.warn}
\title{Prints a warning message.}
\usage{
  log.warn(msg, ...)
}
\arguments{
  \item{msg}{the message that should be printed.}

  \item{...}{any additional text that should be printed.}
}
\description{
  This function will print a warning message.
}
\examples{
\dontrun{
log.warn("detected NA values")
}
}
\author{
  Rob Kooper
}

