% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.debug}
\alias{logger.debug}
\title{Prints a debug message.}
\usage{
logger.debug(msg, ...)
}
\arguments{
\item{msg}{the message that should be printed.}

\item{...}{any additional text that should be printed.}
}
\description{
This function will print a debug message.
}
\examples{
\dontrun{
logger.debug("variable", 5)
}
}
\author{
Rob Kooper
}

