% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logger.error}
\alias{logger.error}
\title{Prints an error message.}
\usage{
logger.error(msg, ...)
}
\arguments{
\item{msg}{the message that should be printed.}

\item{...}{any additional text that should be printed.}
}
\description{
This function will print an error message.
}
\examples{
\dontrun{
logger.error("system did not converge")
}
}
\author{
Rob Kooper
}

