% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logger.getLevel}
\alias{logger.getLevel}
\title{Get configured logging level.}
\usage{
logger.getLevel()
}
\value{
level the level of the message (ALL, DEBUG, INFO, WARN, ERROR, OFF)
}
\description{
This will return the current level configured of the logging messages
}
\examples{
\dontrun{
logger.getLevel()
}
}
\author{
Rob Kooper
}

