% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.getLevelNumber}
\alias{logger.getLevelNumber}
\title{Returns numeric value for string}
\usage{
logger.getLevelNumber(level)
}
\value{
level the level of the message
}
\description{
Given the string representation this will return the numeric value
ALL   =  0
DEBUG = 10
INFO  = 20
WARN  = 30
ERROR = 40
ALL   = 99
}
\author{
Rob Kooper
}

