% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logger.setLevel}
\alias{logger.setLevel}
\title{Configure logging level.}
\usage{
logger.setLevel(level)
}
\arguments{
\item{level}{the level of the message (ALL, DEBUG, INFO, WARN, ERROR, OFF)}
}
\description{
This will configure the logger level. This allows to turn DEBUG, INFO,
WARN and ERROR messages on and off.
}
\examples{
\dontrun{
logger.setLevel("DEBUG")
}
}
\author{
Rob Kooper
}

