% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logger.setOutputFile}
\alias{logger.setOutputFile}
\title{Configure logging output filename.}
\usage{
logger.setOutputFile(filename)
}
\arguments{
\item{filename}{the file to send the log messages to (or NA to not write to file)}
}
\description{
The name of the file where the logging information should be written to.
}
\examples{
\dontrun{
logger.setOutputFile("pecan.log")
}
}
\author{
Rob Kooper
}

