\name{logger.setUseConsole}
\alias{logger.setUseConsole}
\title{Configure logging to console.}
\usage{
  logger.setUseConsole(console)
}
\arguments{
  \item{console}{set to true to print logging to console.}
}
\description{
  Should the logging to be printed to the console or not.
}
\examples{
\dontrun{
logger.setUseConsole(TRUE)
}
}
\author{
  Rob Kooper
}

