% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.setWidth}
\alias{logger.setWidth}
\title{Configure the number of chars per line}
\usage{
logger.setWidth(width)
}
\arguments{
\item{width}{number of chars to print before wrapping to next line.}
}
\description{
The default is for 60 chars per line. Setting this to any value will
wrap the line when printing a message at that many chars.
}
\examples{
\dontrun{
logger.setWidth(70)
}
}
\author{
David LeBauer
}

