% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.warn}
\alias{logger.warn}
\title{Prints a warning message.}
\usage{
logger.warn(msg, ...)
}
\arguments{
\item{msg}{the message that should be printed.}

\item{...}{any additional text that should be printed.}
}
\description{
This function will print a warning message.
}
\examples{
\dontrun{
logger.warn("detected NA values")
}
}
\author{
Rob Kooper
}

