% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.list2init.R
\name{mcmc.list2init}
\alias{mcmc.list2init}
\title{Convert mcmc.list to initial condition list}
\usage{
mcmc.list2init(dat)
}
\arguments{
\item{dat}{mcmc.list object}
}
\value{
list
}
\description{
Used for restarting MCMC code based on last parameters sampled (e.g. in JAGS)
}
\author{
Mike Dietze
}
