% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.utils.R
\name{met2model.exists}
\alias{met2model.exists}
\title{checks that met2model function exists}
\usage{
met2model.exists(model)
}
\arguments{
\item{model}{model package name}
}
\value{
logical
}
\description{
Checks if \verb{met2model.<model>} exists for a particular model
}
