% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{misc.are.convertible}
\alias{misc.are.convertible}
\title{function to check whether units are convertible by misc.convert function}
\usage{
misc.are.convertible(u1, u2)
}
\arguments{
\item{u1}{unit to be converted from, character}

\item{u2}{unit to be converted to, character}
}
\value{
logical
}
\description{
function to check whether units are convertible by misc.convert function
}
\author{
Istem Fer, Shawn Serbin
}
