% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{misc.convert}
\alias{misc.convert}
\title{conversion function for the unit conversions that udunits cannot handle but often needed in PEcAn calculations}
\usage{
misc.convert(x, u1, u2)
}
\arguments{
\item{x}{convertible values}

\item{u1}{unit to be converted from, character}

\item{u2}{unit to be converted to, character}
}
\value{
val converted values
}
\description{
conversion function for the unit conversions that udunits cannot handle but often needed in PEcAn calculations
}
\author{
Istem Fer, Shawn Serbin
}
