\name{mmm2mon}
\alias{mmm2mon}
\title{mmm2mon}
\usage{
  mmm2mon(mmm, lang = "English")
}
\arguments{
  \item{mmm}{three letter character string for month}

  \item{lang}{currently english and portugese}
}
\value{
  month as three letter
}
\description{
  convert month abbrev. to numeric
}
\details{
  Function that determines the number of the month given
  the character name.
}

