% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.output.R
\name{model2netcdfdep}
\alias{model2netcdfdep}
\title{Convert model output to NetCDF}
\usage{
model2netcdfdep(runid, outdir, model, lat, lon, start_date, end_date)
}
\arguments{
\item{runid}{}

\item{model}{name of simulation model currently accepts ("ED2", "SIPNET", "BIOCRO")}

\item{lat}{Latitude of the site}

\item{lon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\value{
vector of filenames created, converts model output to netcdf as a side effect
}
\description{
Convert output for a single model run to NetCDF
}
\details{
DEPRECATED this function will be removed in future versions, please update
your workflow.

This function is a wrapper for model-specific conversion functions,
e.g. \code{model2netcdf.ED2}, \code{model2netcdf.BIOCRO}.
}
\author{
Mike Dietze, David LeBauer
}

