% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mstmipvar}
\alias{mstmipvar}
\title{return MstMIP variable as ncvar}
\usage{
mstmipvar(
  name,
  lat = NULL,
  lon = NULL,
  time = NULL,
  nsoil = NULL,
  silent = FALSE
)
}
\arguments{
\item{name}{of variable}

\item{lat}{latitude if dimension requests it}

\item{lon}{longitude if dimension requests it}

\item{time}{time if dimension requests it}

\item{nsoil}{nsoil if dimension requests it}

\item{silent}{logical: suppress log messages about missing variables?}
}
\value{
ncvar based on MstMIP definition
}
\description{
returns a MstMIP variable as a ncvar based on name and other parameters
passed in.
}
\author{
Rob Kooper
}
