% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{n}
\alias{n}
\title{Summarize Results}
\format{An object of class \code{NULL} of length 0.}
\usage{
summarize.result(result)
}
\arguments{
\item{result}{dataframe with results of trait data query}
}
\value{
result with replicate observations summarized
}
\description{
Summarize results of replicate observations in trait data query
}
\author{
David LeBauer, Alexey Shiklomanov
}
\keyword{datasets}
