% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_leap_day.R
\name{n_leap_day}
\alias{n_leap_day}
\title{n_leap_day}
\usage{
n_leap_day(start_date, end_date)
}
\arguments{
\item{start_date, end_date}{dates in any format recognized by \code{\link[base]{as.Date}}}
}
\description{
number of leap days between two dates
}
\author{
Mike Dietze
}
