% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_model_netcdf.R
\name{nc_merge_all_sites_by_year}
\alias{nc_merge_all_sites_by_year}
\title{Combine many netCDFs into one file per year}
\usage{
nc_merge_all_sites_by_year(
  settings.dir = NULL,
  nc.outdir,
  ancillary.inputs = list(model.outdir = NULL, ens.num = NULL, site.ids = NULL, cores =
    NULL, start.date = NULL, end.date = NULL, time.step = NULL)
)
}
\arguments{
\item{settings.dir}{character: physical path to the pecan standard multi-settings file.}

\item{nc.outdir}{character: physical path to the folder that contains the merged netCDF files.}

\item{ancillary.inputs}{list: necessary arguments if settings.dir is NULL. See
\code{model.outdir} path to the folder that contains model outputs;
\code{ens.num} number of ensembles for the model run;
\code{site.ids} vector of site ids across locations;
\code{cores} number of CPUs for the parallel computation;
\code{start.date} start date of the model run;
\code{end.date} end date of the model run.
\code{time.step} time step of the model run (e.g., 1 year).}
}
\value{
character: file paths to the merged netCDF files.
}
\description{
Merges model outputted netCDF files by the time steps specified in a pecan settings file.
}
\details{
The function is only tested for SIPNET model runs that were run with state data assimilation enabled.
Please make sure you have the same netCDF formats if you want to proceed with different models.
We could also have more functions that deal with different dimensions (e.g., by site instead of by year).
}
\author{
Dongchen Zhang
}
