% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_model_netcdf.R
\name{nc_merge_single_site}
\alias{nc_merge_single_site}
\title{Merge model outputted netCDF files across ensembles for a single site.}
\usage{
nc_merge_single_site(model.outdir, nc.outdir, ens.num, site.id, time)
}
\arguments{
\item{model.outdir}{character: physical path to the model output folder.}

\item{nc.outdir}{character: physical path to the folder that contains the merged netCDF files.}

\item{ens.num}{numeric: ensemble size.}

\item{site.id}{numeric: identification number of the site.}

\item{time}{numeric or character: the current time of netCDF files to be extracted.
See details for use with non-numeric siteIDs}
}
\value{
character: file path to the merged netCDF file.
}
\description{
Merge model outputted netCDF files across ensembles for a single site.
}
\details{
The function is only tested for SIPNET model runs.
Please make sure you have the same netCDF formats if you want to proceed with different models.

This function requires \code{site.id} to be an integer.
If your sites have non-numeric IDs, one possible workaround is to
pass a dummy value and then edit the output file afterward to replace its \code{site_id} variable
with character data. If you do this, do be aware many legacy netCDF tools have poor support
for netCDFs containing character data.
}
\author{
Dongchen Zhang
}
