% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_write_varfiles.R
\name{nc_write_varfiles}
\alias{nc_write_varfiles}
\title{Summarize netcdf variables into text file}
\usage{
nc_write_varfiles(nc_dir, write_mode = c("paired", "collected"))
}
\arguments{
\item{nc_dir}{directory to scan for netCDF files (including in subdirs)}

\item{write_mode}{how many files to write:
"paired" creates files named \emph{.nc.var\verb{alongside every}}.nc\verb{, "collected" writes a single }<nc_dir>/nc_vars.txt` that lists every variable
defined in any nc file in the directory}
}
\description{
Summarize netcdf variables into text file
}
