% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_packages.R
\name{need_packages}
\alias{need_packages}
\title{Check if required packages are installed, and throw an informative
error if not.}
\usage{
need_packages(...)
}
\arguments{
\item{...}{Package names, as characters. Can be passed as
individual arguments, character vectors, or any combination thereof.}
}
\value{
`pkgs`, invisibly
}
\description{
Check if required packages are installed, and throw an informative
error if not.
}
\examples{
# Only need ::: because package isn't exported.
# Inside a package, just call `need_packages`
PEcAn.utils:::need_packages("stats", "methods") # Always works 
try(PEcAn.utils:::need_packages("notapackage"))
}
\author{
Alexey Shiklomanov
}
