% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf2df.R
\name{netcdf2df}
\alias{netcdf2df}
\title{Read all variables from a netCDF into a single data frame}
\usage{
netcdf2df(path)
}
\arguments{
\item{path}{path to a netcdf file}
}
\value{
data frame with columns for each dim and var of the input file.
Units for all of these are attached as an attribute.
}
\description{
Reads all dimensions and variables from a netCDF and returns them as
a single data frame with one row per cell of the source file's
dimension array.
Units are also read and are attached to the result as attribute "units"
}
\details{
Written mostly for files where all dimensions are 1d vectors,
e.g. single-site soil or met files.
Many files with more complex dimensions should work
(at least for cases where it's clear how to rectangle them),
but they are not yet well tested.
}
\author{
Chris Black
}
