% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.tunnel.R
\name{open_tunnel}
\alias{open_tunnel}
\title{Title}
\usage{
open_tunnel(remote_host, tunnel_dir = "~/.pecan/tunnel/", user = NULL,
  password = NULL)
}
\arguments{
\item{remote_host}{name of remote server to connect to (e.g. geo.bu.edu)}

\item{tunnel_dir}{directory to store tunnel file in, typically from settings$host}

\item{user}{username on remote_host}

\item{password}{password on remote_host}
}
\description{
Title
}

