% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{PEcAn.utils-defunct}
\alias{PEcAn.utils-defunct}
\alias{convert.input}
\alias{convert.input-defunct}
\title{Defunct functions in PEcAn.utils}
\description{
The functions listed below are defunct and have been removed from the package.
Calling them will produce a message indicating what function, if any, has replaced it.
}
\section{\code{convert.input}}{
 Use \code{PEcAn.DB::convert_input} instead.
}

\keyword{internal}
