\name{plot.variance.decomposition}
\alias{plot.variance.decomposition}
\title{Plot results of variance decomposition}
\usage{
  plot.variance.decomposition(plot.inputs, outdir,
    prior.plot.inputs = NULL,
    fontsize = list(title = 18, axis = 14), filter = TRUE,
    variance.scale = sqrt, variance.prefix = "Root")
}
\description{
  Plot results of variance decomposition
}
\author{
  David LeBauer
}

