% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{read_jaspar_metadata}
\alias{read_jaspar_metadata}
\title{read_jaspar_metadata}
\usage{
read_jaspar_metadata(file)
}
\arguments{
\item{file}{Pat with JASPAR metadata file.}
}
\value{
Data Frame with JASPAR metadata.
}
\description{
Read file JASPAR 2020 or 2022 metadata of all motifs.
}
