% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_web_config.R
\name{read_web_config}
\alias{read_web_config}
\title{read_web_config}
\usage{
read_web_config(php.config = "../../web/config.php")
}
\arguments{
\item{php.config}{Path to `config.php`}
}
\value{
config.list
}
\description{
read_web_config
}
\author{
Michael Dietze and Rob Kooper
}
