% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regrid}
\alias{regrid}
\title{regrid}
\usage{
regrid(latlon.data)
}
\arguments{
\item{latlon.data}{dataframe with lat, lon, and some value to be regridded}
}
\value{
dataframe with regridded data
}
\description{
Regrid dataset to even grid
}
\author{
David LeBauer
}

