% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{remote.copy.from}
\alias{remote.copy.from}
\title{Copy file from remote to local}
\usage{
remote.copy.from(host, src, dst, delete = FALSE, stderr = FALSE)
}
\arguments{
\item{host}{list with server, user and optionally tunnel to use.}

\item{src}{remote file/dir to copy}

\item{dst}{local file/dir to copy to}

\item{delete}{in case of local dir should all non-existent files be removed}

\item{stderr}{should stderr be returned}
}
\value{
output of command executed
}
\description{
Copy file/dir from remote server to local server
}
\details{
Copies the file/dir from the remote server to the local server. If the dst
is a folder it will copy the file into that folder.
}
\examples{
\dontrun{
  host <- list(name="geo.bu.edu", user="kooper", tunnel="/tmp/geo.tunnel")
  remote.copy.from(host, "/tmp/kooper", "/tmp/geo.tmp", delete=TRUE)
}
}
\author{
Rob Kooper
}

