% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retry.func}
\alias{retry.func}
\title{Retry function X times before stopping in error}
\usage{
retry.func(
  expr,
  isError = function(x) inherits(x, "try-error"),
  maxErrors = 5,
  sleep = 0
)
}
\arguments{
\item{expr}{The function to try running}

\item{isError}{function to use for checking whether to try again.
Must take one argument that contains the result of evaluating \code{expr}
and return TRUE if another retry is needed}

\item{maxErrors}{The number of times to retry the function}

\item{sleep}{How long to wait before retrying the function call}
}
\value{
retval returns the results of the function call
}
\description{
Retry function X times before stopping in error
}
\examples{
\dontrun{
file_url <- paste0(
  "https://thredds.daac.ornl.gov/",
  "thredds/dodsC/ornldaac/1220",
  "/mstmip_driver_global_hd_climate_lwdown_1999_v1.nc4"
)
dap <- retry.func(
  ncdf4::nc_open(file_url),
  maxErrors = 10,
  sleep = 2
)
}

}
\author{
Shawn Serbin <adapted from https://stackoverflow.com/questions/20770497/how-to-retry-a-statement-on-error>
}
