% File src/library/utils/man/RHOME.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2013 R Core Team
% Distributed under GPL 2 or later

\name{RHOME}
\alias{RHOME}
\title{R Home Directory}
\description{
  Returns the location of the \R home directory, which is the root of
  the installed \R tree.
}
\usage{
\special{R RHOME}
}
\keyword{utilities}
