% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.analysis.filenames.r
\name{sensitivity.filename}
\alias{sensitivity.filename}
\title{Generate sensitivity analysis filenames}
\usage{
sensitivity.filename(settings, prefix = "sensitivity.samples",
  suffix = "Rdata", all.var.yr = TRUE, pft = NULL,
  ensemble.id = settings$sensitivity.analysis$ensemble.id,
  variable = settings$sensitivity.analysis$variable,
  start.year = settings$sensitivity.analysis$start.year,
  end.year = settings$sensitivity.analysis$end.year)
}
\value{
a filename
}
\description{
Generate sensitivity analysis filenames
}
\details{
Generally uses values in settings, but can be overwritten for manual uses
}
\author{
Ryan Kelly
}

