% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_terra_options.R, R/utils.R
\name{set_terra_options}
\alias{set_terra_options}
\title{Set Terra Options with Buffer}
\usage{
set_terra_options(
  total_memory_gb,
  total_cores,
  resource_percent,
  priority = "default",
  buffer_percent = 10,
  tempdir = tempdir()
)

set_terra_options(
  total_memory_gb,
  total_cores,
  resource_percent,
  priority = "default",
  buffer_percent = 10,
  tempdir = tempdir()
)
}
\arguments{
\item{total_memory_gb}{Numeric, total system
memory in GB.}

\item{total_cores}{Integer, total number of
cores available.}

\item{resource_percent}{Numeric, percentage
of system resources to allocate for the
process.}

\item{priority}{Character, resource
allocation strategy: "memory", "speed", or
"default". Default is "default".}

\item{buffer_percent}{Numeric, additional
percentage of memory reserved as a buffer to
handle spikes. Default is 10.}

\item{tempdir}{Character, directory for
storing temporary files. Default is
\code{tempdir()}.}
}
\value{
Integer, the number of cores to be
used for parallel processing.

Integer, the number of cores to be
used for parallel processing.
}
\description{
Configure \code{terra} options to control memory
usage, core allocation, and temporary file
storage. This function calculates resource
allocations based on total system memory,
number of cores, resource percentage, and
optional memory buffer for spikes. It sets
\code{memfrac}, \code{memmax}, and \code{tempdir} options
in \code{terra}.

Configure \code{terra} options to control memory
usage, core allocation, and temporary file
storage. This function calculates resource
allocations based on total system memory,
number of cores, resource percentage, and
optional memory buffer for spikes. It sets
\code{memfrac}, \code{memmax}, and \code{tempdir} options
in \code{terra}.
}
\examples{
cores <- set_terra_options_with_buffer(
    total_memory_gb = 128,
    total_cores = 20,
    resource_percent = 50,
    priority = "default",
    buffer_percent = 10,
    tempdir = "D:/r_temp/RtmpWSPfnw"
)
print(cores)

cores <- set_terra_options_with_buffer(
  total_memory_gb = 128,
  total_cores = 20,
  resource_percent = 50,
  priority = "default",
  buffer_percent = 10,
  tempdir = "D:/r_temp/RtmpWSPfnw"
)
print(cores)

}
