% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_model_netcdf.R
\name{single_site_nc_merge}
\alias{single_site_nc_merge}
\title{Merge model outputted netCDF files across ensembles for a single site.}
\usage{
single_site_nc_merge(model.outdir, nc.outdir, ens.num, site.id, lat, lon, time)
}
\arguments{
\item{model.outdir}{character: physical path to the model output folder.}

\item{nc.outdir}{character: physical path to the folder that contains the merged netCDF files.}

\item{ens.num}{numeric: ensemble size.}

\item{site.id}{numeric: identification number of the site.}

\item{lat}{numeric: latitude of the site.}

\item{lon}{numeric: longitude of the site.}
}
\value{
character: file path to the merged netCDF file.
}
\description{
Merge model outputted netCDF files across ensembles for a single site.
}
\details{
The function is only tested for SIPNET model runs.
Please make sure you have the same netCDF formats if you want to proceed with different models.
}
\author{
Dongchen Zhang
}
