% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_case.R, R/utils.R
\name{snake_case}
\alias{snake_case}
\title{Convert Strings to Snake Case}
\usage{
snake_case(x)

snake_case(x)
}
\arguments{
\item{x}{Character vector. The input strings to be converted.}
}
\value{
A character vector with the converted strings in snake_case.

A character vector with the converted strings in snake_case.
}
\description{
Converts a vector of strings to snake_case by inserting underscores
between words and converting all characters to lowercase. It also
replaces spaces with underscores.

Converts a vector of strings to snake_case by inserting underscores
between words and converting all characters to lowercase. It also
replaces spaces with underscores.
}
\examples{
snake_case(c("Normal 1991_2020 AHM", "Normal 1991_2020 bFFP"))
# Returns: c("normal_1991_2020_ahm", "normal_1991_2020_b_ffp")
snake_case(c("Normal 1991_2020 AHM", "Normal 1991_2020 bFFP"))
# Returns: c("normal_1991_2020_ahm", "normal_1991_2020_b_ffp")
}
