% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssh}
\alias{ssh}
\title{R implementation of SSH}
\usage{
ssh(host, ..., args = "")
}
\arguments{
\item{host}{(character) machine to connect to}

\item{...}{Commands to execute. Will be passed as a single quoted string}

\item{args}{futher arguments}
}
\description{
R implementation of SSH
}
