% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.model.runs.R
\name{start.model.runs}
\alias{start.model.runs}
\title{Start ecosystem model runs}
\usage{
\method{start}{model.runs}(settings, write = TRUE, stop.on.error = TRUE)
}
\arguments{
\item{settings}{input pecan settings file}

\item{write}{TRUE/FALSE. Default TRUE}

\item{stop.on.error}{TRUE/FALSE. Default TRUE}
}
\description{
Start selected ecosystem model runs within PEcAn workflow
}
\examples{
\dontrun{
start.model.runs(settings)
}
}
\author{
Shawn Serbin, Rob Kooper, David LeBauer
}
