% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_column_classes.R
\name{summarize_column_classes}
\alias{summarize_column_classes}
\title{Generate a Data Frame of Column Classes}
\usage{
summarize_column_classes(all_data)
}
\arguments{
\item{all_data}{List. A list of data frames.}
}
\value{
A data frame with rows representing the data frames and
columns representing the column names from all the data frames. The
values are the classes of the columns or NA if the column is
missing in a data frame.
}
\description{
This function takes a list of data frames and returns a summary
data frame showing the classes of each column across all the
data frames. Missing columns in some data frames are represented
as NA.
}
\examples{
# Example usage:
all_data <- list(
    df1 = data.frame(a = 1:3, b = letters[1:3]),
    df2 = data.frame(a = 4:6, c = c(TRUE, FALSE, TRUE))
)
column_classes_df <- generate_column_classes(all_data)
print(column_classes_df)

}
