% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{tabnum}
\alias{tabnum}
\title{Table numbers}
\usage{
tabnum(x, n = 3)
}
\arguments{
\item{x}{numeric value or vector}

\item{n}{number of significant figures}
}
\value{
x rounded to n significant figures
}
\description{
Convert number to n significant digits
}
\examples{
tabnum(1.2345)
tabnum(1.2345, n = 4)
}
\author{
David LeBauer
}

