% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test.remote}
\alias{test.remote}
\title{Test Remote}
\usage{
test.remote(host)
}
\arguments{
\item{host}{}
}
\value{
logical - TRUE if remote connection is available
}
\description{
Test ssh access
}
\details{
Test to determine if access to a remote server is available.
Can be used to exclude / include tests or to prevent / identify access errors
}
\author{
Rob Kooper
}

