% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_science.R, R/utils.R
\name{theme_science}
\alias{theme_science}
\title{Minimal Theme for Scientific Plots}
\usage{
theme_science()

theme_science()
}
\value{
A ggplot2 theme object.

A ggplot2 theme object.
}
\description{
This function creates minimalistic ggplot2 theme for scientific
plots in ggplot2. It removes gridlines, simplifies axis
formatting, and adds clean, bold titles and labels.

This function creates minimalistic ggplot2 theme for scientific
plots in ggplot2. It removes gridlines, simplifies axis
formatting, and adds clean, bold titles and labels.
}
