% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_nc.R
\name{to_ncvar}
\alias{to_ncvar}
\title{Define an NCDF variable}
\usage{
to_ncvar(varname, dims)
}
\arguments{
\item{varname}{character vector, standard variable name (must be in PEcAn.utils::standard_vars)}

\item{dims}{list of previously defined ncdims (function will match subset of dims for this variable in standard_vars; can include other dims--enables lapply.)}
}
\value{
ncvar defined according to standard_vars
}
\description{
Define an NCDF variable
}
\author{
Anne Thomas
}
