% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryl}
\alias{tryl}
\title{Test if function gives an error}
\usage{
tryl(FUN)
}
\arguments{
\item{FUN}{function to be evaluated for error}
}
\value{
FALSE if function returns error; else TRUE
}
\description{
adaptation of try that returns a logical value (FALSE if error)
}
\examples{
tryl(1 + 1)
# TRUE
tryl(sum("a"))
# FALSE
}
\author{
David LeBauer
}
