% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_convert.R
\name{ud_convert}
\alias{ud_convert}
\title{Convert units

This function takes the numeric argument \code{x}, quantified in units \code{u1} and converts it to be of units \code{u2}.}
\usage{
ud_convert(x, u1, u2)
}
\arguments{
\item{x}{numeric vector with units \code{u1} to be converted to units \code{u2}}

\item{u1}{string which is parseable into a \code{units} compatible unit.}

\item{u2}{string which is parseable into a \code{units} compatible unit,
and for which there exists a defined transformation from the units represented by \code{u1}.}
}
\value{
numeric vector stripped of units
}
\description{
Convert units

This function takes the numeric argument \code{x}, quantified in units \code{u1} and converts it to be of units \code{u2}.
}
\author{
Chris Black
}
