% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_are_equivalent.R
\name{units_are_equivalent}
\alias{units_are_equivalent}
\title{Check if two unit strings are equivalent}
\usage{
units_are_equivalent(x, y)
}
\arguments{
\item{x}{A unit string, as character}

\item{y}{Another unit string for comparison, as character}
}
\value{
\code{TRUE} if equivalent, \code{FALSE} otherwise
}
\description{
This is to allow multiple forms of the same unit to work, such as
\code{m/s} vs. \verb{m s-1} or \code{K} and \code{Kelvin}.
}
\author{
Alexey Shiklomanov
}
