% File src/library/utils/man/upgrade.Rd
% Part of the R package, https://www.R-project.org
% Copyright 2023 R Core Team
% Distributed under GPL 2 or later

\name{upgrade}
\alias{upgrade}
\title{Upgrade}
\description{Upgrade objects.}
\usage{
upgrade(object, \dots)
}
\arguments{
  \item{object}{an \R object.}
  \item{\dots}{further arguments passed to or from other methods.}
}
\details{
  This is a generic function, with a method for
  \code{"\link{packageStatus}"} objects.
}
