\name{utils-package}
\alias{utils-package}
\alias{utils}
\docType{package}
\title{
The R Utils Package
}
\description{
R utility functions
}
\details{
This package contains a collection of utility functions.

For a complete
list, use \code{library(help="utils")}.

}
\author{
R Development Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ programming }
