% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vecpaste}
\alias{vecpaste}
\title{Convert vector to comma delimited string}
\usage{
vecpaste(x)
}
\arguments{
\item{x}{vector}
}
\value{
comma delimited string
}
\description{
vecpaste, turns vector into comma delimited string fit for SQL statements.
}
