% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{write_jaspar_metadata}
\alias{write_jaspar_metadata}
\title{write_jaspar_metadata}
\usage{
write_jaspar_metadata(metadata)
}
\arguments{
\item{metadata}{Data Frame with JASPAR metadata.}
}
\value{
Write file meta_jaspar.txt in the current directory.
}
\description{
Write file JASPAR 2020 or 2022 metadata of all motifs.
}
