% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero.bounded.density}
\alias{zero.bounded.density}
\title{Zero bounded density using log density transform}
\usage{
zero.bounded.density(x, bw = "SJ", n = 1001)
}
\arguments{
\item{x}{data, as a numeric vector}

\item{bw}{The smoothing bandwidth to be used. See 'bw.nrd'}

\item{n}{number of points to use in kernel density estimate. See \code{\link[stats]{density}}}
}
\value{
data frame with back-transformed log density estimate
}
\description{
Provides a zero bounded density estimate of a parameter.
Kernel Density Estimation used by the \code{\link[stats]{density}} function will cause problems
at the left hand end because it will put some weight on negative values.
One useful approach is to transform to logs, estimate the density using KDE, and then transform back.
}
\references{
M. P. Wand, J. S. Marron and D. Ruppert, 1991. Transformations in Density Estimation. Journal of the American Statistical Association. 86(414):343-353 \url{http://www.jstor.org/stable/2290569}
}
\author{
\href{https://stats.stackexchange.com/q/6588/2750}{Rob Hyndman}
}
