% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{zero.truncate}
\alias{zero.truncate}
\title{Zero Truncate}
\usage{
zero.truncate(y)
}
\arguments{
\item{y}{numeric vector}
}
\value{
numeric vector with all values less than 0 set to 0
}
\description{
Truncates vector at 0
}
\author{
<unknown>
}

