\name{bibtexify}
\alias{bibtexify}
\title{bibtexify}
\usage{
bibtexify(author, year, title)
}
\arguments{
  \item{author}{name of first author}

  \item{year}{year of publication}

  \item{title}{manuscript title}
}
\value{
bibtex citation
}
\description{
Convert author, year, title to bibtex citation format
}
\details{
Converts author year title to author1999abc format
}

