% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.scratch.R
\name{clear.scratch}
\alias{clear.scratch}
\title{Removes previous model run output from worker node local scratch directories on EBI-CLUSTER}
\usage{
clear.scratch(settings)
}
\arguments{
\item{settings}{list of PEcAn settings. Only \code{settings$host$name} is used}
}
\value{
nothing
}
\description{
Removes previous model run output from worker node local scratch directories on EBI-CLUSTER
}
\examples{
\dontrun{
clear.scratch(settings)
}
}
\author{
Shawn Serbin
}
