% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clear.scratch}
\alias{clear.scratch}
\title{Clear EBI-CLUSTER worker node local scratch directories of old PEcAn output}
\usage{
clear.scratch(settings)
}
\value{
nothing
}
\description{
Removes previous model run output from worker node local scratch directories on EBI-CLUSTER
}
\examples{
\dontrun{
clear.scratch(settings)
}
}
\author{
Shawn Serbin
}

