% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert.expr}
\alias{convert.expr}
\title{Convert expression to variable names}
\usage{
convert.expr(expression)
}
\arguments{
\item{expression}{expression string}
}
\value{
list
}
\description{
Convert expression to variable names
}
\author{
Istem Fer
}
