% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.input.R
\name{convert.input}
\alias{convert.input}
\title{convert.input}
\usage{
convert.input(input.id, outfolder, formatname, mimetype, site.id, start_date,
  end_date, pkg, fcn, username, con = con, hostname = "localhost", browndog,
  write = TRUE, ...)
}
\description{
Convert input by applying fcn and insert new record into database
}
\author{
Betsy Cowdery, Michael Dietze
}

