% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.input}
\alias{convert.input}
\title{Convert input by applying fcn and insert new record into database}
\usage{
convert.input(input.id, outfolder, pkg, fcn, write, username, con, ...)
}
\description{
Convert input by applying fcn and insert new record into database
}

