% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf2date.R
\name{datetime2doy}
\alias{datetime2doy}
\alias{cf2doy}
\title{Extract Julian day from CF or POSIXct date-times}
\usage{
datetime2doy(datetime, tz = "UTC")

cf2doy(value, unit, tz = "UTC")
}
\arguments{
\item{datetime}{POSIXct datetime, or object that can be to POSIXct
via \code{as.POSIXct}}

\item{tz}{Time zone of result (default = "UTC")}

\item{value}{Numeric value of CF date-time}

\item{unit}{CF style unit (e.g. "days since 2010-01-01")}
}
\value{
Numeric Julian date
}
\description{
This gets around the fact that most functions for calculating
Julian Day do not support non-integer days.
}
\examples{
datetime2doy("2010-01-01") # 1
datetime2doy("2010-01-01 12:00:00") # 1.5
cf2doy(0, "days since 2007-01-01") 
cf2doy(5, "days since 2010-01-01") # 6
cf2doy(5, "days since 2010-01-01") # 6
}
\author{
Alexey Shiklomanov

Alexey Shiklomanov
}
