% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_year.R
\name{days_in_year}
\alias{days_in_year}
\title{Number of days in a year}
\usage{
days_in_year(year, leap_year = TRUE)
}
\arguments{
\item{year}{Numeric year (can be a vector)}

\item{leap_year}{Default = TRUE. If set to FALSE will always return 365}
}
\value{
integer vector, all either 365 or 366
}
\description{
Calculate number of days in a year based on whether it is a leap year or not.
}
\examples{
days_in_year(2010)  # Not a leap year -- returns 365
days_in_year(2012)  # Leap year -- returns 366
days_in_year(2000:2008)  # Function is vectorized over years
}
\author{
Alexey Shiklomanov
}
