\name{distn.table.stats}
\alias{distn.table.stats}
\title{return mean and standard deviation of a distribution for each distribution in a table with \code{colnames = c("distn", "a", "b"), e.g. in a table of priors}
\usage{
distn.table.stats(distns)
}
\arguments{
  \item{distns}{table of distributions; see examples}
}
\value{
named vector of mean and SD
}
\description{
a helper function for computing summary statistics of a
parametric distribution
}
\author{
David LeBauer
}

