% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_conversions.R
\name{do_conversions}
\alias{do_conversions}
\alias{do.conversions}
\title{do_conversions}
\usage{
do_conversions(settings, overwrite.met = FALSE, overwrite.fia = FALSE,
  overwrite.ic = FALSE)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{overwrite.met, overwrite.fia, overwrite.ic}{logical}
}
\description{
Input conversion workflow

DEPRECATED: This function has been moved to the PEcAn.workflow package and will be removed from PEcAn.utils.
}
\author{
Ryan Kelly, Rob Kooper, Betsy Cowdery, Istem Fer
}
