% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download.file}
\alias{download.file}
\title{download.file}
\usage{
download.file(url, filename, method)
}
\arguments{
\item{url}{complete URL for file download}

\item{filename}{destination file name}

\item{method}{Method of file retrieval. Can set this using the options(download.ftp.method=\link{method}) in your Rprofile.
example options(download.ftp.method="ncftpget")}
}
\description{
Simple function to use ncftpget for FTP downloads behind a firewall.
Requires ncftpget and a properly formatted config file in the users
home directory
}
\examples{
\dontrun{
download.file("http://lib.stat.cmu.edu/datasets/csb/ch11b.txt","~/test.download.txt")

download.file("
  ftp://ftp.cdc.noaa.gov/Datasets/NARR/monolevel/pres.sfc.2000.nc",
  "~/pres.sfc.2000.nc")
}

}
\author{
Shawn Serbin, Rob Kooper
}
