% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.url.R
\name{download.url}
\alias{download.url}
\title{Download file from the url.}
\usage{
download.url(url, file, timeout = 600, .opts = list(), retry404 = TRUE)
}
\arguments{
\item{timeout}{number of seconds to wait for file (default 600)}

\item{retry404}{retry on a 404, this is used by Brown Dog}

\item{the}{url of the file to download}

\item{the}{filename}

\item{list}{of options for curl, for example to download from a
protected site use list(userpwd=userpass, httpauth = 1L)}
}
\value{
returns name of file if successfull or NA if not.
}
\description{
Try and download a file.
}
\details{
This will download a file, if retry404 and 404 is returned it will
wait until the file is available. If the file is still not available
after timeout tries, it will return NA. If the file is downloaded
it will return the name of the file
}
\examples{
\dontrun{
download.url('http://localhost/', index.html)
}
}

