\name{flush.console}
\alias{flush.console}
\title{Flush Output to A Console}
\description{
  This does nothing except on console-based versions of \R.
  On the Mac OS X and Windows GUIs, it ensures that the display of
  output in the console is current, even if output buffering is on.
}
\usage{
flush.console()
}

\keyword{ utilities }
