% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full.path.R
\name{full.path}
\alias{full.path}
\title{Creates an absolute path to a folder.}
\usage{
full.path(folder)
}
\arguments{
\item{folder}{folder for file paths.}
}
\value{
absolute path
}
\description{
This will take a folder and make it into an absolute folder name. It
will normalize the path and prepend it with the current working folder
if needed to get an absolute path name.
}
\examples{
full.path('pecan')
}
\author{
Rob Kooper
}
