\name{get.model.output}
\alias{get.model.output}
\title{Retrieve model output}
<<<<<<< HEAD
\usage{
  get.model.output(model, settings)
}
=======
>>>>>>> ef9572c0bfd74f64f90e160d3d29f5e3f933b0b6
\arguments{
  \item{model}{the ecosystem model run}
}
\description{
  This function retrieves model output for further analyses
}
\examples{
\dontrun{
get.model.output(model)
get.model.output("ED2")
}
}
\author{
  Michael Dietze, Shawn Serbin, David LeBauer
}

