% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.write.configs.R
\name{get.parameter.samples}
\alias{get.parameter.samples}
\title{Sample from priors or posteriors}
\usage{
get.parameter.samples(pfts = settings$pfts)
}
\arguments{
\item{pfts}{the pfts node of the list of pecan settings}
}
\description{
Convert priors / MCMC samples to chains that can be sampled for model parameters
}
\author{
David LeBauer, Shawn Serbin
}

