% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.parameter.stat}
\alias{get.parameter.stat}
\title{Get Parameter Statistics}
\usage{
get.parameter.stat(mcmc.summary, parameter)
}
\arguments{
\item{mcmc.summary}{}

\item{parameter}{}
}
\value{
table with parameter statistics
}
\description{
Gets statistics for LaTeX - formatted table
}
\examples{
\dontrun{get.parameter.stat(mcmc.summaries[[1]], 'beta.o')}
}
\author{
David LeBauer
}

