\name{get.run.id}
\alias{get.run.id}
\title{Get Run ID}
\usage{
  get.run.id(run.type, index, trait = "", pft.name = "")
}
\arguments{
  \item{run.type}{}

  \item{index}{}

  \item{trait}{}

  \item{pft.name}{}
}
\value{
  id representing a model run
}
\description{
  returns an id representing a model run
}
\details{
  for use in model input files and indices
}

