% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.config.utils.R
\name{get.sa.sample.list}
\alias{get.sa.sample.list}
\title{get.sa.sample.list}
\usage{
get.sa.sample.list(pft, env, quantiles)
}
\arguments{
\item{pft}{Plant Functional Type}

\item{env}{}

\item{quantiles}{quantiles at which to obtain samples from parameter for
sensitivity analysis}
}
\value{
sa.sample.list
}
\description{
get sensitivity samples as a list
}

