\name{get.stats.mcmc}
\alias{get.stats.mcmc}
\title{Get stats for parameters in MCMC output}
\usage{
  get.stats.mcmc(mcmc.summary, sample.size)
}
\arguments{
  \item{mcmc.summary}{}

  \item{sample.size}{}
}
\value{
  list with summary statistics for parameters in an MCMC
  chain
}
\description{
  Further summarizes output from summary.mcmc
}
\author{
  David LeBauer
}

