% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid.R
\name{grid2netcdf}
\alias{grid2netcdf}
\title{grid2netcdf}
\usage{
grid2netcdf(gdata, date = "9999-09-09", outfile = "out.nc")
}
\arguments{
\item{grid.data}{}
}
\value{
nothing, writes netCDF file
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
David LeBauer
}

