% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listToXml}
\alias{listToXml}
\title{List to XML}
\usage{
listToXml(item, tag)
}
\arguments{
\item{item}{}

\item{tag}{xml tag}
}
\value{
xmlNode
}
\description{
Convert List to XML
}
\details{
Can convert list or other object to an xml object using xmlNode
}
\author{
unknown
}

