\name{load.modelpkg}
\alias{load.modelpkg}
\title{Load model package}
\usage{
load.modelpkg(model)
}
\arguments{
  \item{model}{name of model}
}
\value{
FALSE if function returns error; else TRUE
}
\description{
load model package
}
\examples{
\dontrun{require.modelpkg(BioCro)}
}
\author{
David LeBauer
}

