% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdata_files.R, R/utils.R
\name{load_rdata_files}
\alias{load_rdata_files}
\title{Load .RData Files into a List}
\usage{
load_rdata_files(data_dir)

load_rdata_files(data_dir)
}
\arguments{
\item{data_dir}{Character. The path to the directory containing
\code{.RData} files.}
}
\value{
A list where each element is an object loaded from the
\code{.RData} files.

A list where each element is an object loaded from the
\code{.RData} files.
}
\description{
This function scans a directory for \code{.RData} files, loads each file,
and stores all objects from the \code{.RData} files into a single list.

This function scans a directory for \code{.RData} files, loads each file,
and stores all objects from the \code{.RData} files into a single list.
}
