% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.info}
\alias{logger.info}
\title{Prints an informational message.}
\usage{
logger.info(msg, ...)
}
\arguments{
\item{msg}{the message that should be printed.}

\item{...}{any additional text that should be printed.}
}
\description{
This function will print an informational message.
}
\examples{
\dontrun{
logger.info("PEcAn version 1.2")
}
}
\author{
Rob Kooper
}

