% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.setOutputFile}
\alias{logger.setOutputFile}
\title{Configure logging output filename.}
\usage{
logger.setOutputFile(filename)
}
\arguments{
\item{filename}{the file to send the log messages to (or NA to not write to file)}
}
\description{
The name of the file where the logging information should be written to.
}
\examples{
\dontrun{
logger.setOutputFile("pecan.log")
}
}
\author{
Rob Kooper
}

