% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logger.severe}
\alias{logger.severe}
\title{Prints an severe message and stops execution.}
\usage{
logger.severe(msg, ...)
}
\arguments{
\item{msg}{the message that should be printed.}

\item{...}{any additional text that should be printed.}
}
\description{
This function will print a message and stop execution of the code. This
should only be used if the application should terminate.
}
\details{
set \code{\link{logger.setQuitOnSevere(FALSE)}}. To avoid terminating
the session. This is set by default to TRUE if interactive or running
inside Rstudio.
}
\examples{
\dontrun{
logger.severe("missing parameters")
}
}
\author{
Rob Kooper
}

