% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.config.utils.R
\name{met2model.exists}
\alias{met2model.exists}
\title{met2model.exists}
\usage{
met2model.exists(model)
}
\arguments{
\item{model}{model package name}
}
\value{
logical
}
\description{
checks that met2model function exists
}
\details{
Checks if met2model.<model> exists for a particular
model
}

