% Generated by roxygen2 (4.0.2): do not edit by hand
\name{met2model.exists}
\alias{met2model.exists}
\title{met2model.exists}
\usage{
met2model.exists(model)
}
\arguments{
\item{model}{model package name}
}
\value{
logical
}
\description{
checks that met2model function exists
}
\details{
Checks if met2model.<model> exists for a particular
model
}

