% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{mstmipvar}
\alias{mstmipvar}
\title{MstMIP variable}
\usage{
mstmipvar(name, lat = NA, lon = NA, time = NA, nsoil = NA,
  silent = FALSE)
}
\arguments{
\item{name}{name of variable}

\item{lat}{latitude if dimension requests it}

\item{lon}{longitude if dimension requests it}

\item{time}{time if dimension requests it}

\item{nsoil}{nsoil if dimension requests it}
}
\value{
ncvar based on MstMIP definition
}
\description{
return MstMIP variable as ncvar
}
\details{
returns a MstMIP variable as a ncvar based on name and other parameters
passed in.
}
\author{
Rob Kooper
}

