% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{newxtable}
\alias{newxtable}
\title{newxtable}
\usage{
newxtable(x, environment = "table", table.placement = "ht", label = NULL,
  caption = NULL, caption.placement = NULL, align = NULL)
}
\arguments{
\item{x}{data.frame to be converted to latex table}

\item{environment}{can be 'table'; 'sidewaystable' if using latex rotating package}

\item{table.placement}{}

\item{label}{}

\item{caption}{}

\item{caption.placement}{}

\item{align}{}
}
\value{
Latex version of table, with percentages properly formatted
}
\description{
New xtable
}
\details{
utility to properly escape the "%" sign for latex
}
\author{
David LeBauer
}

