\name{paste.stats}
\alias{paste.stats}
\title{Paste Stats}
\usage{
  paste.stats(mcmc.summary, median, lcl, ucl, n = 2)
}
\arguments{
  \item{mcmc.summary}{}

  \item{median}{}

  \item{lcl}{}

  \item{ucl}{}

  \item{n}{}
}
\description{
  A helper function for building a LaTex table.
}
\details{
  Used by \code{\link{get.parameter.stat}}.
}
\author{
  David LeBauer
}

