% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{read.ensemble.output}
\alias{read.ensemble.output}
\title{Read ensemble output}
\usage{
read.ensemble.output(ensemble.size, pecandir, outdir, start.year, end.year,
  variable, ens.run.ids = NULL)
}
\arguments{
\item{ensemble.size}{the number of ensemble members run}

\item{pecandir}{specifies where pecan writes its configuration files}

\item{outdir}{directory with model output to use in ensemble analysis}

\item{start.year}{first year to include in ensemble analysis}

\item{end.year}{last year to include in ensemble analysis}

\item{variables}{targe variables for ensemble analysis}
}
\value{
a list of ensemble model output
}
\description{
Reads output from model ensemble
}
\details{
Reads output for an ensemble of length specified by \code{ensemble.size} and bounded by \code{start.year}
and \code{end.year}
}
\author{
Ryan Kelly, David LeBauer, Rob Kooper
}

