% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{read.sa.output}
\alias{read.sa.output}
\title{Read Sensitivity Analysis output}
\usage{
read.sa.output(traits, quantiles, pecandir, outdir, pft.name = "", start.year,
  end.year, variable, sa.run.ids = NULL)
}
\arguments{
\item{traits}{model parameters included in the sensitivity analysis}

\item{quantiles}{quantiles selected for sensitivity analysis}

\item{pecandir}{specifies where pecan writes its configuration files}

\item{outdir}{directory with model output to use in sensitivity analysis}

\item{pft.name}{name of PFT used in sensitivity analysis (Optional)}

\item{start.year}{first year to include in sensitivity analysis}

\item{end.year}{last year to include in sensitivity analysis}

\item{variables}{variables to be read from model output}
}
\value{
dataframe with one col per quantile analysed and one row per trait,
 each cell is a list of AGB over time
}
\description{
Reads output of sensitivity analysis runs
}
\author{
Ryan Kelly, David LeBauer, Rob Kooper, Mike Dietze
}

