% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{regrid}
\usage{
regrid(latlon.data)
}
\arguments{
\item{latlon.data}{dataframe with lat, lon, and some value to be regridded}
}
\value{
dataframe with regridded data
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
David LeBauer
}

