\name{run.write.configs}
\alias{run.write.configs}
\title{Run model specific write configuration functions}
\usage{
  run.write.configs(model)
}
\arguments{
  \item{model}{the ecosystem model to generate the
  configuration files for}
}
\description{
  Main driver function to call the ecosystem model specific
  (e.g. ED, SiPNET) run and configuration file scripts
}
\author{
  David LeBauer, Shawn Serbin
}

