% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seconds_in_year.R
\name{seconds_in_year}
\alias{seconds_in_year}
\title{Number of seconds in a given year}
\usage{
seconds_in_year(year, leap_year = TRUE)
}
\arguments{
\item{year}{Numeric year (can be a vector)}

\item{leap_year}{Default = TRUE. If set to FALSE will always return 31536000}
}
\description{
Number of seconds in a given year
}
\examples{
seconds_in_year(2000)  # Leap year -- 366 x 24 x 60 x 60 = 31622400
seconds_in_year(2001)  # Regular year -- 365 x 24 x 60 x 60 = 31536000
seconds_in_year(2000:2005)  # Vectorized over year
}
\author{
Alexey Shiklomanov
}
