% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{sendmail}
\alias{sendmail}
\title{Sends email. This assumes the program sendmail is installed.}
\usage{
sendmail(from, to, subject, body)
}
\arguments{
\item{from}{the sender of the mail message}

\item{to}{the receipient of the mail message}

\item{subject}{the subject of the mail message}

\item{body}{the body of the mail message}
}
\value{
nothing
}
\description{
Sends email. This assumes the program sendmail is installed.
}
\examples{
\dontrun{
sendmail('bob@example.com', 'joe@example.com', 'Hi', 'This is R.')
}
}
\author{
Rob Kooper
}
