\name{start.model.runs}
\alias{start.model.runs}
\title{Start ecosystem model runs}
<<<<<<< HEAD
\usage{
  start.model.runs(model, write = TRUE)
}
=======
>>>>>>> ef9572c0bfd74f64f90e160d3d29f5e3f933b0b6
\description{
  Start selected ecosystem model runs within PEcAn workflow
}
\examples{
\dontrun {
start.model.runs("ED2")
start.model.runs("SIPNET")
}
}
\author{
  Shawn Serbin
}

