% Generated by roxygen2 (4.0.1): do not edit by hand
\name{start.model.runs}
\alias{start.model.runs}
\title{Start ecosystem model runs}
\usage{
\method{start}{model.runs}(settings, write = TRUE)
}
\description{
Start selected ecosystem model runs within PEcAn workflow
}
\examples{
\dontrun{
start.model.runs("ED2")
start.model.runs("SIPNET")
}
}
\author{
Shawn Serbin, Rob Kooper, ...
}

