\name{start.model.runs}
\alias{start.model.runs}
\title{Start ecosystem model runs}
\usage{
\method{start}{model.runs}(model, write = TRUE)
}
\description{
Start selected ecosystem model runs within PEcAn workflow
}
\examples{
\dontrun {
start.model.runs("ED2")
start.model.runs("SIPNET")
}
}
\author{
Shawn Serbin, Rob Kooper, ...
}

