\name{summarize.result}
\alias{summarize.result}
\title{Summarize Results}
\usage{
summarize.result(result)
}
\arguments{
  \item{result}{dataframe with results of trait data query}
}
\value{
result with replicate observations summarized
}
\description{
Summarize results of replicate observations in trait data
query
}
\author{
David LeBauer
}

