% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone_hour.R
\name{timezone_hour}
\alias{timezone_hour}
\title{Timezone Hour}
\usage{
timezone_hour(timezone)
}
\arguments{
\item{timezone}{to be converted}
}
\value{
hours offset of the timezone
}
\description{
Returns the number of hours offset to UTC for a timezone.
}
\examples{
\dontrun{
timezone_hour('America/New_York')
}
}
\author{
Rob Kooper
}
