% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_nc.R
\name{to_ncdim}
\alias{to_ncdim}
\title{Make some values into an NCDF dimension variable}
\usage{
to_ncdim(dimname, vals)
}
\arguments{
\item{dimname}{character vector, standard dimension name (must be in PEcAn.utils::standard_vars)}

\item{vals}{values of dimension; can be single value or vector}
}
\value{
ncdim defined according to standard_vars
}
\description{
Units and longnames are looked up from the \code{\link{standard_vars}} table
}
\author{
Anne Thomas
}
