% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transformstats}
\alias{transformstats}
\title{Transform Stats}
\usage{
transformstats(data)
}
\arguments{
\item{data}{data frame with mean, statistic, n, and statistic name: \code{example data <- data.frame(Y=rep(1,5), stat=rep(1,5), n=rep(4,5), statname=c('SD', 'MSE', 'LSD', 'HSD', 'MSD'))}}
}
\value{
dataframe with statistics transformed to SE
}
\description{
Transform misc. statistics to SE
}
\details{
Automates transformations of SD, MSE, LSD, 95\%CI, HSD, and MSD to conservative estimates of SE.
}
\examples{
statdf <- data.frame(Y=rep(1,5),
                               stat=rep(1,5),
                               n=rep(4,5),
                               statname=c('SD', 'MSE', 'LSD', 'HSD', 'MSD'))
transformstats(statdf)
}
\author{
David LeBauer
}

