% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_convert.R
\name{ud_convert}
\alias{ud_convert}
\title{Convert units}
\usage{
ud_convert(x, u1, u2)
}
\arguments{
\item{x}{vector of class "numeric" or "difftime"}

\item{u1}{string parseable as the units in which \code{x} is provided.  If \code{x} is
class "difftime", then \code{u1} is not actually used.  However, it still needs
to be supplied and needs to be convertible to \code{u2} for consistency.}

\item{u2}{string parseable as the units to convert to}
}
\value{
numeric vector with values converted to units in \code{u2}
}
\description{
Unit conversion to replace the now-unmaintained \code{udunits2::ud.convert}
}
\author{
Chris Black
}
