% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_is_parseable.R
\name{unit_is_parseable}
\alias{unit_is_parseable}
\title{Check whether a unit string is parseable}
\usage{
unit_is_parseable(u1)
}
\arguments{
\item{u1}{A character string representing a type of units}
}
\value{
TRUE if the units is parseable, FALSE otherwise.
}
\description{
Function will replace the now-unmaintained \code{udunits2::ud.is.parseable}
}
\author{
Tanishq Jain
}
