\name{utils-defunct}
%------ NOTE:  ../R/utils-defunct.R   must be synchronized with this!
\alias{utils-defunct}
\alias{package.contents}
\title{Defunct Functions in Package utils}
\description{
  The functions or variables listed here are no longer part of \R as
  they are not needed (any more).
}
\usage{
package.contents(pkg, lib.loc = NULL)
}
\seealso{
  \code{\link{utils-deprecated}},
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{internal}
