% File src/library/utils/man/utils-deprecated.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2013 R Core Team
% Distributed under GPL 2 or later

\name{utils-deprecated}
%------ NOTE:  ../R/utils-deprecated.R   must be synchronized with this!
\alias{utils-deprecated}
\title{Deprecated Functions in Package \pkg{utils}}
\description{
  These functions are provided for compatibility with older versions of
  \R only, and may be defunct as soon as of the next release.
}
\seealso{
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{misc}
