% File src/library/utils/man/utils-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{utils-package}
\alias{utils-package}
\alias{utils}
\docType{package}
\title{
The R Utils Package
}
\description{
R utility functions
}
\details{
This package contains a collection of utility functions.

For a complete
list, use \code{library(help = "utils")}.

}
\author{
R Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ programming }
