% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vecpaste}
\alias{vecpaste}
\title{vecpaste}
\usage{
vecpaste(x)
}
\arguments{
\item{x}{vector}
}
\value{
comma delimited string
}
\description{
Convert vector to comma delimited string
}
\details{
vecpaste, turns vector into comma delimited string fit for SQL statements.
}

