% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{write.ensemble.configs}
\alias{write.ensemble.configs}
\title{Write ensemble configs}
\usage{
write.ensemble.configs(defaults, ensemble.samples, settings, model,
  clean = FALSE, write.to.db = TRUE, restart = NULL)
}
\arguments{
\item{defaults}{pft}

\item{ensemble.samples}{list of lists supplied by \link{get.ensemble.samples}}

\item{settings}{list of PEcAn settings}

\item{clean}{remove old output first?}

\item{restart}{In case this is a continuation of an old simulation. restart needs to be a list with name tags of runid, inputs, new.params (parameters), new.state (initial condition), ensemble.id (ensemble ids), start.time and stop.time.}

\item{write.config}{a model-specific function to write config files, e.g. \link{write.config.ED}}
}
\value{
list, containing $runs = data frame of runids, and $ensemble.id = the ensemble ID for these runs. Also writes sensitivity analysis configuration files as a side effect
}
\description{
Write ensemble config files
}
\details{
Writes config files for use in meta-analysis and returns a list of run ids.
Given a pft.xml object, a list of lists as supplied by get.sa.samples, 
a name to distinguish the output files, and the directory to place the files.
}
\author{
David LeBauer, Carl Davidson, Hamze Dokoohaki
}
