% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{write.sa.configs}
\alias{write.sa.configs}
\title{Write sensitivity analysis config files}
\usage{
write.sa.configs(defaults, quantile.samples, settings, model,
  clean = FALSE, write.to.db = TRUE)
}
\arguments{
\item{defaults}{named list with default parameter values}

\item{quantile.samples}{list of lists supplied by \link{get.sa.samples}}

\item{settings}{list of settings}

\item{model}{name of model to be run}

\item{clean}{logical: Delete any existing contents of the directory specified by \code{settings$rundir} before writing to it?}

\item{write.to.db}{logical: Record this run to BETY? If TRUE, uses connection settings specified in \code{settings$database}}
}
\value{
list, containing $runs = data frame of runids, and $ensemble.id = the ensemble ID for these runs. Also writes sensitivity analysis configuration files as a side effect
}
\description{
Writes config files for use in sensitivity analysis.
}
\author{
David LeBauer, Carl Davidson
}
